pyproject.toml
[project]
name = "vpcsc-mcp"
version = "0.1.0"
description = "MCP server for VPC Service Controls setup and implementation — helps AI agents and developers manage GCP VPC-SC perimeters, access levels, and troubleshooting"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "ONS Cloud Team" },
]
dependencies = [
    "mcp[cli]>=1.26.0,<2",
    "pydantic>=2.12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
adk = [
    "google-adk>=1.27.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vpcsc_mcp"]

[project.scripts]
vpcsc-mcp = "vpcsc_mcp.server:main"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
